Add service worker support (#101)
Allow service worker options to be specified in the dojorc. If the file
path for an existing service worker is provided, it is copied to the
output directory. Otherwise, a service worker file is generated that
contains a precache manifest, as well as any specified route handlers.

Further, when generating a service worker from the rc, an additional
entry point is added to the build that initializes the service worker at
startup.

Finally, add PWA-specific meta tags to the application's index.html
(requires an upgraded to html-webpack-plugin).